requirements.txt
Django==2.0.2
django-braces==1.12.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-model-utils==3.1.1
django-registration-redux==2.1
django-webpack-loader==0.5.0
djangorestframework==3.7.7
psycopg2==2.7.4
pytz==2017.3
sqlparse==0.2.4
